begin modularization
